+++ .babelrc
@@ -1,8 +1,3 @@
 {
-  "plugins": ["transform-es2015-modules-umd"],
-  "env": {
-    "travis": {
-      "plugins": [ "__coverage__" ]
-    }
-  }
+  "presets": ["env"]
 }
