+++ .gitignore
@@ -1,85 +1,32 @@
-#####=== Linux ===#####
 *~
-
-# KDE directory preferences
 .directory
-
-# Linux trash folder which might appear on any partition or disk
 .Trash-*
-
-#####=== Node ===#####
-
-# Logs
 logs
 *.log
-
-# Runtime data
 pids
 *.pid
 *.seed
-
-# Generated test bundle
-test/browser/test_bundle.js
-
-# Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
-
-# Coverage directory used by tools like istanbul
-test/coverage/**/*
-
-# Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)
-.grunt
-
-# node-waf configuration
 .lock-wscript
-
-# Compiled binary addons (http://nodejs.org/api/addons.html)
 build/Release
-
-# Dependency directory
-# https://www.npmjs.org/doc/misc/npm-faq.html#should-i-check-my-node_modules-folder-into-git-
 node_modules
-
-# Debug log from npm
 npm-debug.log
 npm-debug.log*
-
-#####=== OSX ===#####
 .DS_Store
 .AppleDouble
 .LSOverride
-
-# Icon must end with two \r
 Icon
-
-
-# Thumbnails
 ._*
-
-# Files that might appear on external disk
 .Spotlight-V100
 .Trashes
-
-# Directories potentially created on remote AFP share
 .AppleDB
 .AppleDesktop
 Network Trash Folder
 Temporary Items
 .apdisk
-
-#####=== SublimeText ===#####
-# cache files for sublime text
 *.tmlanguage.cache
 *.tmPreferences.cache
 *.stTheme.cache
-
-# workspace files are user-specific
 *.sublime-workspace
-
-# project files should be checked into the repository, unless a significant
-# proportion of contributors will probably not be using SublimeText
-# *.sublime-project
-
-# sftp configuration file
 sftp-config.json
-
+coverage
