+++ package.json
@@ -5,15 +5,10 @@
   "main": "src/index.js",
   "scripts": {
     "build": "make build && git changelog",
-    "codecov": "make codecov",
     "readme": "node create-readme.js",
-    "test": "make test",
-    "test:browser": "make test-browser",
-    "test:karma": "make test-karma",
-    "test:karma:watch": "make test-karma-watch",
-    "test:travis": "make test-travis",
-    "test:node": "babel-node --presets es2015 test/runners/node.js",
-    "semantic-release": "make semantic-release",
+    "test": "jest",
+    "test:ci": "jest --runInBand --coverage",
+    "test:watch": "jest --watch",
     "amend": "git add . && git commit --amend --reuse-message=HEAD",
     "push": "git push --tags && git push",
     "release": "npm version patch && make build && npm run amend && npm run push && npm publish"
@@ -29,38 +24,15 @@
   },
   "homepage": "https://github.com/devlucky/Kakapo.js#readme",
   "devDependencies": {
-    "babel-core": "^5.4.3",
-    "babel-loader": "5.4.0",
-    "babel-plugin-__coverage__": "0.111111.0",
-    "babel-plugin-transform-es2015-modules-umd": "^6.8.0",
-    "babel-preset-es2015": "6.6.0",
-    "babelify": "7.3.0",
-    "browserify": "13.0.0",
-    "codecov.io": "0.1.6",
-    "eslint": "2.9.0",
-    "eslint-config-airbnb-base": "2.0.0",
-    "eslint-plugin-import": "1.7.0",
-    "ghooks": "1.2.1",
+    "babel-jest": "^22.4.3",
+    "babel-preset-env": "^1.6.1",
     "http-server": "~0.7.1",
+    "jest": "^22.4.3",
     "jquery": "^2.2.4",
     "json-loader": "0.5.4",
-    "karma": "0.13.22",
-    "karma-browserify": "5.0.3",
-    "karma-coverage": "0.5.5",
-    "karma-firefox-launcher": "0.1.7",
-    "karma-tap": "1.0.4",
-    "mocha-phantomjs": "^4.0.1",
     "npm-run-all": "1.7.0",
-    "semantic-release": "4.3.5",
     "serve": "1.4.0",
-    "superagent": "^2.0.0",
-    "tap-spec": "^4.1.1",
-    "tape": "^4.5.1",
-    "tape-dom": "0.0.10",
-    "tape-run": "^2.1.3",
-    "watchify": "3.7.0",
-    "webpack": "~1.4.4",
-    "webpack-dev-server": "~1.6.5"
+    "superagent": "^2.0.0"
   },
   "dependencies": {
     "faker": "^3.1.0",
@@ -69,21 +41,6 @@
     "path-match": "1.2.4",
     "path-to-regexp": "^1.2.1",
     "query-string": "4.1.0"
-  },
-  "browserify": {
-    "transform": [
-      [
-        "babelify",
-        {
-          "presets": [
-            "es2015"
-          ]
-        }
-      ]
-    ]
-  },
-  "directories": {
-    "test": "test"
   },
   "keywords": [
     "mocking",
