--- test/browser/fixtures.json
@@ -1,6 +1,0 @@
-{
-  "user_ids": [1, 2, 3],
-  "comments": [{
-    "id": 1
-  }]
-}
