--- test/browser/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Kakapo.js - tests</title>
-</head>
-<body>
-  <script src="test_bundle.js"></script>
-</body>
-</html>
