--- test/runners/node.js
@@ -1,3 +1,0 @@
-import * as specs from '../specs/node/';
-
-Object.keys(specs).forEach(specName => specs[specName]());
