--- test/runners/travis.js
@@ -1,3 +1,0 @@
-import * as specs from '../specs';
-
-Object.keys(specs).forEach(specName => specs[specName]());
