--- test/specs/index.js
@@ -1,7 +1,0 @@
-export { databaseSpec } from './database_spec';
-export { routerSpec } from './router_spec';
-export { serverSpec } from './server_spec';
-export { responseSpec } from './response_spec';
-export { requestSpec } from './request_spec';
-export { serializerSpec } from './serializer_spec';
-export { relationshipsSpec } from './relationships_spec';
