--- test/specs/integration/index.js
@@ -1,2 +1,0 @@
-export { jquerySpec } from './jquery_spec';
-export { superagentSpec } from './superagent_spec';
