--- test/specs/node/index.js
@@ -1,1 +1,0 @@
-export { routerSpec } from './router_spec';
