--- test/specs/node/router_spec.js
@@ -1,15 +1,0 @@
-import test from 'tape';
-import { Router } from '../../../src';
-
-export const routerSpec = () => {
-  //TODO: Test get request using options as string
-  test('Node # Router # get string', (assert) => {
-
-  });
-
-  //TODO: Test get request using options as object
-  //TODO: Test different options properties
-  test('Node # Router # get options', (assert) => {
-
-  });
-}
